Add dart_code_metrics
