+++ places/lib/ui/components/components.dart
@@ -1,2 +1,2 @@
 export 'btn_secondary.dart';
-export 'btn_square.dart';
+export 'btn_white_square.dart';
