+++ places/lib/ui/components/label_text_widget.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:places/res/res.dart';
 
-class LabelWidget extends StatelessWidget {
+class LabelTextWidget extends StatelessWidget {
   final String text;
 
-  const LabelWidget(this.text, {Key? key}) : super(key: key);
+  const LabelTextWidget(this.text, {Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
