+++ places/pubspec.yaml
@@ -14,11 +14,13 @@ dependencies:
     sdk: flutter
   flutter_svg: ^0.21.0+1
   provider: ^5.0.0
-  surf_lint_rules: ^1.3.0
 
 dev_dependencies:
+  dart_code_metrics: ^4.2.0-dev.1
   flutter_test:
     sdk: flutter
+  surf_lint_rules: ^1.3.0
+
 flutter:
   uses-material-design: true
   assets:
